+++ cbits/HsGLUT.c
@@ -48,12 +48,6 @@ hs_GLUT_getProcAddress(const char *name)
 #include <stdlib.h>
 #include <dlfcn.h>
 
-#ifdef __APPLE__
-#define FILENAME "/System/Library/Frameworks/GLUT.framework/GLUT"
-#else
-#define FILENAME "libglut.so"
-#endif
-
 void*
 hs_GLUT_getProcAddress(const char *name)
 {
@@ -62,7 +56,18 @@ hs_GLUT_getProcAddress(const char *name)
 
   if (firstTime) {
     firstTime = 0;
-    handle = dlopen(FILENAME, RTLD_LAZY | RTLD_GLOBAL);
+
+#ifdef __APPLE__
+    /* Try public framework path first. */
+    handle = dlopen("/Library/Frameworks/GLUT.framework/GLUT", RTLD_LAZY | RTLD_GLOBAL);
+
+    /* If the public path failed, try the system framework path. */
+    if (!handle) {
+      handle = dlopen("/System/Library/Frameworks/GLUT.framework/GLUT", RTLD_LAZY | RTLD_GLOBAL);
+    }
+#else
+    handle = dlopen("libglut.so", RTLD_LAZY | RTLD_GLOBAL);
+#endif
   }
 
   return handle ? dlsym(handle, name) : NULL;